Add new env variables and fix frontend
HealthCheckButton was not rendered in frontend
Also change ChatbotService to use newly renamed env variable.
