+++ frontend/src/App.jsx
@@ -1,39 +1,39 @@
 import { useState, useEffect } from 'react';
-import axios from 'axios';
+// import axios from 'axios';
 
 const App = () => {
   // Luodaan kaksi react statea
-  const [input, setInput] = useState('');
-  const [messages, setMessages] = useState([]);
+  // const [input, setInput] = useState('');
+  // const [messages, setMessages] = useState([]);
 
-  useEffect(() => {
-    fetch('http://double-agent-backend:8000/messages')
-      .then((res) => res.json())
-      .then((data) => setMessages(data.messages.map((text) => ({ role: 'user', content: text }))));
-  }, []); // useEffect hook hakee viestit backendistä vain kerran, kun komponentti mountataan eli kun sivu ladataan
+  // useEffect(() => {
+  //   fetch('http://double-agent-backend:8000/messages')
+  //     .then((res) => res.json())
+  //     .then((data) => setMessages(data.messages.map((text) => ({ role: 'user', content: text }))));
+  // }, []); // useEffect hook hakee viestit backendistä vain kerran, kun komponentti mountataan eli kun sivu ladataan
 
-  const handleSubmit = async (e) => {
-    e.preventDefault(); // Tämä estää html form elementin oletuskäyttäytymisen (sivun uudelleenlataus)
-    try {
-      const res = await axios.post('http://double-agent-backend:8000/chat', {
-        message: input,
-        thread_id: 'default',
-      }); // Lähetetään POST pyyntö backendille axios kirjaston avulla
+  // const handleSubmit = async (e) => {
+  //   e.preventDefault(); // Tämä estää html form elementin oletuskäyttäytymisen (sivun uudelleenlataus)
+  //   try {
+  //     const res = await axios.post('http://double-agent-backend:8000/chat', {
+  //       message: input,
+  //       thread_id: 'default',
+  //     }); // Lähetetään POST pyyntö backendille axios kirjaston avulla
 
-      setMessages((prev) => [
-        ...prev,
-        { role: 'user', content: res.data.user_message },
-        { role: 'ai', content: res.data.ai_response },
-      ]);
-      setInput(''); // Tyhjennetään input kenttä
-    } catch (err) {
-      console.error(err);
-    }
-  };
+  //     setMessages((prev) => [
+  //       ...prev,
+  //       { role: 'user', content: res.data.user_message },
+  //       { role: 'ai', content: res.data.ai_response },
+  //     ]);
+  //     setInput(''); // Tyhjennetään input kenttä
+  //   } catch (err) {
+  //     console.error(err);
+  //   }
+  // };
 
   return (
     <div style={{ padding: '2rem' }}>
-      <h1>Our simple chatbot</h1>
+      {/* <h1>Our simple chatbot</h1>
       <form onSubmit={handleSubmit}>
         <input
           value={input}
@@ -59,7 +59,32 @@ const App = () => {
             </div>
           ))}
         </div>
-      </div>
+      </div> */}
+      <h1>Chatti pois käytöstä kunnes yliopistokirjautuminen on kunnossa.</h1>
+      <HealthCheckButton />
+    </div>
+  );
+};
+
+export const HealthCheckButton = () => {
+  const [status, setStatus] = useState(null);
+
+  const checkHealth = async () => {
+    // try {
+    //   const res = await fetch('http://double-agent-backend:8000/health');
+    //   const data = await res.json();
+    //   setStatus(data.status);
+    // } catch (err) {
+    //   console.error(err);
+    //   setStatus('error');
+    // }
+    console.log('Health check is currently disabled.');
+  };
+
+  return (
+    <div>
+      <button onClick={checkHealth}>Check Backend Health</button>
+      {status && <p>Backend status: {status}</p>}
     </div>
   );
 };
